pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs", "hatch-fancy-pypi-readme"]
build-backend = "hatchling.build"

[project]
name = "investai"
version = "0.1.0"
description = "AI-powered stock investment analysis platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "InvestAI Team", email = "team@investai.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    # Core Dependencies
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    # Database
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "alembic==1.12.1",
    "redis==5.0.1",
    # AI/ML
    "langchain==0.0.350",
    "openai==1.3.5",
    "sentence-transformers==2.2.2",
    "faiss-cpu==1.7.4",
    "numpy==1.24.4",
    "pandas==2.0.3",
    "scikit-learn==1.3.2",
    # Data Processing
    "aiohttp==3.9.1",
    "requests==2.31.0",
    "beautifulsoup4==4.12.2",
    "lxml==4.9.3",
    # Utilities
    "python-dotenv==1.0.0",
    "python-dateutil==2.8.2",
    "pytz==2023.3.post1",
    "loguru==0.7.2",
    "pyjwt==2.8.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.2.0",
    # Trading Integration (Optional)
    "msgpack==1.0.7",
    "protobuf==4.25.1",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    
    # Code quality
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
    "mypy==1.7.0",
    "pre-commit==3.5.0",
    
    # Documentation
    "mkdocs==1.5.3",
    "mkdocs-material==9.5.3",
    "mkdocstrings[python]==0.23.0",
    
    # Development tools
    "ipython==8.18.1",
    "jupyter==1.0.0",
    
    # Monitoring
    "prometheus-client==0.19.0",
    
    # API testing
    "httpx==0.25.2",
    
    # Database tools
    "sqlalchemy-utils==0.41.1",
]

[project.urls]
Homepage = "https://github.com/your-org/InvestAI"
Documentation = "https://investai.readthedocs.io"
Repository = "https://github.com/your-org/InvestAI"
Issues = "https://github.com/your-org/InvestAI/issues"


[tool.hatch.build]
include = [
    "app/**/*",
    "tests/**/*",
    "README.md",
    "LICENSE",
    "pyproject.toml",
]

[tool.hatch.version]
source = "vcs"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
show_error_codes = true
strict = true

[tool.pytest.ini_options]
addopts = "-v --tb=short -x"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "B", "C", "N", "UP", "YTT"]
ignore = []

[tool.ruff.lint]
# Enable pycodestyle (E, W) and pyflakes (F) rules
select = ["E", "F", "W"]

[tool.ruff.format]
# Use Black-compatible formatting
quote-style = "double"

# Package discovery
[tool.setuptools.packages.find]
where = ["app"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml", "*.yml"]
